requirements.txt
Flask==3.0.0
pg8000==1.30.5
python-dotenv==1.0.0